feat: Extract the interfaces in separate file and create folder for every component
