+++ jeopardy-FE/src/components/Column/Column.tsx
@@ -1,22 +1,14 @@
-import {
-  Dispatch, SetStateAction, useEffect, useState,
-} from 'react';
+import { useEffect, useState } from 'react';
 import axios from 'axios';
-import Flipper from './Flipper';
+import Flipper from '../Flipper/Flipper';
 import styles from './Column.module.scss';
-import { IQuestion } from '../sharedInterfaces';
+import IColumn from './IColumn';
+import apiInstance from '../../services/axiosConfig';
 
 interface IProps {
   id: number;
   actualQuestion: string;
   score: number;
-}
-
-interface IColumn {
-  category: string;
-  setActive: Dispatch<SetStateAction<boolean>>;
-  setIsInputBlocked: Dispatch<SetStateAction<boolean>>;
-  setSelectedQuestion: Dispatch<SetStateAction<IQuestion | undefined>>;
 }
 
 function Column({
@@ -27,10 +19,8 @@ function Column({
 }: IColumn) {
   const [questions, setQuestions] = useState([]);
   useEffect(() => {
-    axios.defaults.headers.post['Content-Type'] = 'application/json;charset=utf-8';
-    axios.defaults.headers.post['Access-Control-Allow-Origin'] = '*';
-    axios
-      .get(`http://localhost:8080/question/${category}`)
+    apiInstance
+      .get(`/question/${category}`)
       .then((response) => setQuestions(response.data));
   }, [category]);
   const column: JSX.Element[] = [];
