+++ jeopardy-FE/src/components/Flipper/Flipper.module.scss
