+++ jeopardy-FE/src/components/Input/IInput.ts
@@ -1,0 +1,8 @@
+export default interface IInput {
+  value?: string;
+  handleChange: (e: React.FormEvent<HTMLInputElement>) => void;
+  label: string;
+  placeholder: string;
+  text: string;
+  disabled?: boolean;
+}
