+++ jeopardy-FE/src/components/Main/Main.module.scss
