+++ jeopardy-FE/src/components/NumberInput/NumberInput.module.scss
