+++ jeopardy-FE/src/components/OptionsItem/OptionsItem.tsx
