+++ jeopardy-FE/src/components/QuestionsMatrix/IQuestionsMatrix.ts
@@ -1,0 +1,9 @@
+import { Dispatch, SetStateAction } from 'react';
+import { IQuestion } from '../../sharedInterfaces';
+
+export default interface IQuestionsMatrix {
+  categories: string[];
+  setActive: Dispatch<SetStateAction<boolean>>;
+  setSelectedQuestion: Dispatch<SetStateAction<IQuestion | undefined>>;
+  setIsInputBlocked: Dispatch<SetStateAction<boolean>>;
+}
