+++ jeopardy-FE/src/components/QuestionsMatrix/QuestionsMatrix.tsx
@@ -1,21 +1,14 @@
 /* eslint-disable react/no-array-index-key */
-import { Dispatch, SetStateAction } from 'react';
-import Column from './Column';
+import Column from '../Column/Column';
 import styles from './QuestionsMatrix.module.scss';
-import { IQuestion } from '../sharedInterfaces';
+import IQuestionsMatrix from './IQuestionsMatrix';
 
-interface IMatrix {
-  categories: string[];
-  setActive: Dispatch<SetStateAction<boolean>>;
-  setSelectedQuestion: Dispatch<SetStateAction<IQuestion | undefined>>;
-  setIsInputBlocked: Dispatch<SetStateAction<boolean>>;
-}
 export default function Matrix({
   categories,
   setActive,
   setSelectedQuestion,
   setIsInputBlocked,
-}: IMatrix) {
+}: IQuestionsMatrix) {
   return (
     <div className={styles.questions}>
       {categories.map((category: string, index: number) => (
